requirements.txt
click
cerberus
gitpython
PyYAML
ujson
